+++ modules/kubernetes/aks-core/variables.tf
@@ -469,5 +469,5 @@ variable "azure_metrics_config" {
 variable "node_local_dns_enabled" {
   description = "Should VPA be enabled"
   type        = bool
-  default     = false
+  default     = true
 }